elevia-clinic-onboarder/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
playwright
anthropic
python-dotenv
sqlalchemy
psycopg2-binary
chromadb
